requirements.txt
ctransformers==0.2.27
sentence-transformers==2.5.1
pinecone-client==3.1.0
langchain==0.1.12
langchain-community==0.0.28
flask==3.0.2
pypdf==4.1.0
python-dotenv==1.0.1
-e .